+++ docs/docs/auto-docs/check-i18n.test-utils/functions/cleanupTempDirs.md
@@ -1,0 +1,13 @@
+[Admin Docs](/)
+
+***
+
+# Function: cleanupTempDirs()
+
+> **cleanupTempDirs**(): `void`
+
+Defined in: [src/check-i18n.test-utils.js:40](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/check-i18n.test-utils.js#L40)
+
+## Returns
+
+`void`
