+++ docs/docs/auto-docs/check-i18n.test-utils/functions/makeTempDir.md
@@ -1,0 +1,13 @@
+[Admin Docs](/)
+
+***
+
+# Function: makeTempDir()
+
+> **makeTempDir**(): `string`
+
+Defined in: [src/check-i18n.test-utils.js:27](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/check-i18n.test-utils.js#L27)
+
+## Returns
+
+`string`
