+++ docs/docs/auto-docs/check-i18n.test-utils/functions/runScript.md
@@ -1,0 +1,21 @@
+[Admin Docs](/)
+
+***
+
+# Function: runScript()
+
+> **runScript**(`targets`, `options`): `SpawnSyncReturns`\<`string`\>
+
+Defined in: [src/check-i18n.test-utils.js:15](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/check-i18n.test-utils.js#L15)
+
+## Parameters
+
+### targets
+
+`any`
+
+### options
+
+## Returns
+
+`SpawnSyncReturns`\<`string`\>
