+++ docs/docs/auto-docs/check-i18n.test-utils/functions/writeTempFile.md
@@ -1,0 +1,27 @@
+[Admin Docs](/)
+
+***
+
+# Function: writeTempFile()
+
+> **writeTempFile**(`dir`, `relPath`, `content`): `string`
+
+Defined in: [src/check-i18n.test-utils.js:33](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/check-i18n.test-utils.js#L33)
+
+## Parameters
+
+### dir
+
+`any`
+
+### relPath
+
+`any`
+
+### content
+
+`any`
+
+## Returns
+
+`string`
