+++ docs/docs/auto-docs/check-i18n.test-utils/variables/fixturesDir.md
@@ -1,0 +1,9 @@
+[Admin Docs](/)
+
+***
+
+# Variable: fixturesDir
+
+> `const` **fixturesDir**: `string`
+
+Defined in: [src/check-i18n.test-utils.js:11](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/check-i18n.test-utils.js#L11)
