+++ docs/docs/auto-docs/check-i18n.test-utils/variables/scriptPath.md
@@ -1,0 +1,9 @@
+[Admin Docs](/)
+
+***
+
+# Variable: scriptPath
+
+> `const` **scriptPath**: `string`
+
+Defined in: [src/check-i18n.test-utils.js:10](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/check-i18n.test-utils.js#L10)
